+++ src/main/java/com/jf/PetApp/application/investment/dto/InvestmentLotDTO.java
@@ -2,17 +2,23 @@ package com.jf.PetApp.application.investment.dto;
 
 import com.jf.PetApp.core.domain.enums.InvestmentType;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
+/**
+ * Money and quantity are {@link BigDecimal} — never {@code double} — because
+ * this DTO carries the real_portfolio ledger chain achievement thresholds key
+ * off: see docs/BACKEND_MODULE_PLAN.md §12.
+ */
 public record InvestmentLotDTO(
     Integer id,
     String name,
     InvestmentType type,
-    Double quantity,
-    Double purchasePrice,
+    BigDecimal quantity,
+    BigDecimal purchasePrice,
     LocalDate purchaseDate,
-    Double currentPrice,
-    Double investedValue,
-    Double currentValue
+    BigDecimal currentPrice,
+    BigDecimal investedValue,
+    BigDecimal currentValue
 ) {
 }
