+++ src/main/java/com/jf/PetApp/application/investment/dto/PortfolioHistoryPointDTO.java
@@ -1,10 +1,14 @@
 package com.jf.PetApp.application.investment.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
+/**
+ * Money fields are {@link BigDecimal} — see docs/BACKEND_MODULE_PLAN.md §12.
+ */
 public record PortfolioHistoryPointDTO(
     LocalDate date,
-    Double investedCapital,
-    Double portfolioValue
+    BigDecimal investedCapital,
+    BigDecimal portfolioValue
 ) {
 }
