+++ src/main/java/com/jf/PetApp/application/investment/dto/UserPositionDTO.java
@@ -1,18 +1,23 @@
 package com.jf.PetApp.application.investment.dto;
 
+import java.math.BigDecimal;
+
 /**
  * The authenticated user's aggregated position in a specific asset —
  * computed from their real {@code jf_investments} lots and the current
  * market price. Every field traces back to validated data; nothing is
  * estimated or fabricated.
+ *
+ * <p>Money and quantity are {@link BigDecimal} — see
+ * docs/BACKEND_MODULE_PLAN.md §12.</p>
  */
 public record UserPositionDTO(
-    double quantity,
-    double averagePrice,
-    double investedValue,
-    double currentValue,
-    double unrealizedGain,
-    double unrealizedGainPercent,
-    double portfolioWeight
+    BigDecimal quantity,
+    BigDecimal averagePrice,
+    BigDecimal investedValue,
+    BigDecimal currentValue,
+    BigDecimal unrealizedGain,
+    BigDecimal unrealizedGainPercent,
+    BigDecimal portfolioWeight
 ) {
 }
